refactor(renderer): split App.tsx phase 2, giant hooks, AgentNodeCard
挙動を変えずに、Issue #373 Phase 1 後に肥大が残った 4 領域を責務別へ分割する
(Refs #487)。公開 API・公開シンボル・IPC コマンド名・event 名・三点同期はすべて
不変。React の購読粒度や effect 依存も元の挙動を維持する。

(a) App.tsx Phase 2
- 冒頭の `const settings = { ... useSettingsValue(...) ... }` 合成ブロックを
  `lib/hooks/use-app-shell-state.ts` に切り出し。
- プロジェクトメニュー / ワークスペース系 handler (handleNewProject /
  handleOpenFolder / handleOpenFile / handleOpenRecent / handleClearRecent /
  handleAddWorkspaceFolder / handleRemoveWorkspaceFolder + workspaceFolders memo)
  を `lib/hooks/use-project-loader.ts` に移管。editor タブとの相互作用は
  `discardEditorTabsForRoot` ブリッジで App 側に残し、Issue #33 の「Cancel 時
  に settings / tabs どちらも変更しない」約束を維持する。

(b) AgentNodeCard.tsx (664 行) → folder split
- `cards/AgentNodeCard/CardFrame.tsx` ロール視覚化 / handoff UI / 起動引数解決
- `cards/AgentNodeCard/TerminalOverlay.tsx` PTY 配線 / 入力 auto-summary /
  sessionId 書き戻し / Canvas zoom fit
- `cards/AgentNodeCard/index.tsx` 旧 `default export` 互換層
- 共有型は `cards/AgentNodeCard/types.ts` に集約

(c) use-pty-session.ts (788 行) → 2 hook + 互換 wrapper
- `lib/hooks/use-xterm-bind.ts` 初回接続 / subscribeEventReady pre-subscribe /
  attach 経路の queue→replay→flush / IME composition gate
- `lib/hooks/use-hmr-recover.ts` HMR 判定 (hmrDisposeArmed) / 再接続キャッシュ
  (acquireGeneration / isCurrentGeneration / cacheUpsert / cacheDelete / cacheGet)
- `lib/use-pty-session.ts` は両 hook を組み立てる薄い wrapper。
  `usePtySession` / `PtySpawnSnapshot` / `PtySessionCallbacks` /
  `UsePtySessionOptions` の公開シグネチャは不変。

(d) use-team-management.ts (524 行) → 3 hook + 互換 wrapper
- `lib/hooks/use-team-state.ts` teams / spawnStaggerTimers / TeamHub 接続情報 /
  doCloseTeam / handleCloseLeaderOnly / プロジェクト切替 reset
- `lib/hooks/use-team-history-sync.ts` teamHistoryEntries / debounce save /
  handleResumeTeam / handleDeleteTeamHistory / handleTerminalSessionId /
  persistTerminalCustomLabel + unmount 即 flush + spawn timer 停止
- `lib/hooks/use-team-launch-helpers.ts` getTerminalArgs / getCodexInstructions /
  getRolePrompt / getTerminalEnv
- `lib/hooks/use-team-management.ts` は 3 hook を組み立てる薄い wrapper。
  `UseTeamManagementOptions` / `UseTeamManagementResult` の公開シグネチャは不変。

検証
- `npm run typecheck` green
- `npx vitest run` 250/250 green (use-pty-session-fonts / -hmr / canvas /
  team-prompts-liveness / 他の renderer test を全件再走)

Refs #487
