+++ src/renderer/src/components/canvas/cards/AgentNodeCard/CardFrame.tsx
@@ -1,69 +1,45 @@
 /**
- * AgentNodeCard — チームメンバー (claude/codex のロール持ち端末) を表すカード。
+ * AgentNodeCard / CardFrame
  *
- * TerminalCard の派生で、視覚要素を強化:
- *   - ヘッダー左にロール色のアバター (1 文字 + 背景円)
- *   - 枠線/接続点の色をロール色に揃える
- *   - ステータスバッジ (idle/thinking/typing) を将来用に右上に配置 (Phase 3+)
+ * Issue #487: AgentNodeCard 単一ファイルから「カード枠 + role visual + handoff UI」
+ * を切り出したもの。pty / xterm の配線は隣接の TerminalOverlay.tsx に分けてある。
  *
- * payload: TerminalPayload + role 必須
+ * 責務:
+ *   - NodeResizer + 入出力 Handle (xyflow)
+ *   - ロール由来の accent / avatar / 表示ラベル (resolveAgentVisual)
+ *   - ヘッダー (title / organization / role / StatusBadge / handoff button / close)
+ *   - Leader 専用の handoff 作成 → bracketed paste 注入フロー
+ *   - 起動引数 (command / args / sysPrompt / codexInstructions) の解決
+ *   - TerminalOverlay の mount & 共有 ref / callback の橋渡し
+ *
+ * 挙動は元 AgentNodeCard.tsx と完全一致。構造のみ整理。
  */
-import { memo, useCallback, useEffect, useMemo, useRef, useState } from 'react';
+import { memo, useCallback, useMemo, useRef, useState } from 'react';
 import { Handle, NodeResizer, Position, type NodeProps } from '@xyflow/react';
 import { ClipboardCheck } from 'lucide-react';
-import { TerminalView, type TerminalViewHandle } from '../../TerminalView';
-import { useT } from '../../../lib/i18n';
-import { useSettings } from '../../../lib/settings-context';
-import { useCanvasStore, NODE_MIN_W, NODE_MIN_H } from '../../../stores/canvas';
-import { useCanvasTerminalFit } from '../../../lib/use-canvas-terminal-fit';
-import { useConfirmRemoveCard } from '../../../lib/use-confirm-remove-card';
-import { useXtermScrollToBottomOnResize } from '../../../lib/use-xterm-scroll-on-resize';
-import { renderSystemPrompt, useRoleProfiles } from '../../../lib/role-profiles-context';
-import { resolveAgentVisual } from '../../../lib/agent-visual';
-import { parseShellArgs } from '../../../lib/parse-args';
-import { resolveAgentConfig } from '../../../lib/agent-resolver';
-import { useRecruitSpawnAck } from '../../../lib/use-terminal-spawn';
-import { useToast } from '../../../lib/toast-context';
+import { useT } from '../../../../lib/i18n';
+import { useSettings } from '../../../../lib/settings-context';
+import {
+  useCanvasStore,
+  NODE_MIN_W,
+  NODE_MIN_H
+} from '../../../../stores/canvas';
+import { useConfirmRemoveCard } from '../../../../lib/use-confirm-remove-card';
+import {
+  renderSystemPrompt,
+  useRoleProfiles
+} from '../../../../lib/role-profiles-context';
+import { resolveAgentVisual } from '../../../../lib/agent-visual';
+import { parseShellArgs } from '../../../../lib/parse-args';
+import { resolveAgentConfig } from '../../../../lib/agent-resolver';
+import { useToast } from '../../../../lib/toast-context';
+import type { TerminalViewHandle } from '../../../TerminalView';
 import type {
   HandoffCheckpoint,
-  HandoffReference,
-  TeamOrganizationMeta
-} from '../../../../../types/shared';
-
-interface AgentPayload {
-  agent?: 'claude' | 'codex';
-  /** 新スキーマ: ロール識別子。未設定時は legacy `role` をフォールバックとして読む。 */
-  roleProfileId?: string;
-  /** @deprecated 旧フィールド。canvas store v2 マイグレーションで roleProfileId に移行済み */
-  role?: string;
-  teamId?: string;
-  agentId?: string;
-  command?: string;
-  args?: string[];
-  cwd?: string;
-  /** Claude Code のセッション id。検出時に payload に書き戻し、次回 spawn で
-   *  `--resume <id>` を付与して前回会話を復元する。 */
-  resumeSessionId?: string | null;
-  /**
-   * Issue #117: team_recruit の custom_instructions が新規エージェントに渡るように、
-   * use-recruit-listener.ts が payload に積んでくる「役職追加指示の生テキスト」。
-   *   - Claude  : sysPrompt の末尾に追記して --append-system-prompt に流す。
-   *   - Codex   : codex_instructions として一時ファイル化し、起動時に PTY 注入される。
-   *   - 動的ロール (instructions ベース) と併用された場合は両方をブレンドする。
-   * undefined / 空文字なら「指定なし」と同じ扱い。
-   */
-  customInstructions?: string;
-  /** @deprecated `customInstructions` の旧名。互換のため受理だけする (後方互換)。 */
-  codexInstructions?: string;
-  /** Issue #359: handoff から新セッションを起動するときに初手で送るプロンプト。 */
-  initialMessage?: string;
-  /** Issue #370: 複数組織同時運用時の所属表示・履歴復元用情報。 */
-  organization?: TeamOrganizationMeta;
-  /** Issue #359: 本文はファイル保存し、payload には最新 handoff 参照だけ残す。 */
-  latestHandoff?: HandoffReference;
-}
-
-type AgentStatus = 'idle' | 'thinking' | 'typing';
+  HandoffReference
+} from '../../../../../../types/shared';
+import type { AgentPayload, AgentStatus } from './types';
+import { TerminalOverlay } from './TerminalOverlay';
 
 /**
  * pty 起動時の status 文字列 ("実行中: claude --append-system-prompt ...long text...") を
@@ -77,29 +53,6 @@ function shortStatus(s: string): string {
 }
 
 /**
- * デフォルト (auto-summary 上書き対象) のタイトルかを判定。
- * "Claude #1" / "Codex #2" / "Leader" 等は上書き OK、ユーザーが手で付けた名前は守る。
- */
-function isAutoTitle(t: string): boolean {
-  return /^(Claude|Codex|Agent|Leader|Planner|Programmer|Researcher|Reviewer)( #\d+)?$/i.test(
-    t.trim()
-  );
-}
-
-/** ユーザー入力テキストから「機能追加」のような短いタイトルを抽出 */
-function summarizeInput(text: string): string {
-  const cleaned = text
-    .replace(/[\r\n]+/g, ' ')
-    .replace(/\s+/g, ' ')
-    .trim();
-  if (cleaned.length === 0) return '';
-  // 30 文字で切り詰め、句読点で更に切る
-  const cut = cleaned.slice(0, 30);
-  const punct = cut.search(/[。．、,]/);
-  return punct > 4 ? cut.slice(0, punct) : cut;
-}
-
-/**
  * 絶対パスからファイル名だけを返す。Windows (`\`) と POSIX (`/`) の両方に対応するため
  * path モジュールを使わず手元で処理する (renderer 側に node:path は無い)。
  */
@@ -109,7 +62,9 @@ function basenameOf(absPath: string): string {
   return idx >= 0 ? normalized.slice(idx + 1) : normalized;
 }
 
-function handoffReferenceOf(handoff: HandoffCheckpoint | HandoffReference): HandoffReference {
+function handoffReferenceOf(
+  handoff: HandoffCheckpoint | HandoffReference
+): HandoffReference {
   return {
     id: handoff.id,
     kind: handoff.kind,
@@ -159,21 +114,12 @@ function wrapBracketedPaste(text: string): string {
 }
 
 function AgentNodeCardImpl({ id, data }: NodeProps): JSX.Element {
-  const ref = useRef<TerminalViewHandle | null>(null);
-  // Issue #261: NodeResizer でカードを縮めたあと再度広げたとき、内部 `.xterm-viewport`
-  // の scrollTop が中途半端な位置で残って「末尾が見えない」状態になることがある。
-  // `.canvas-agent-card__term` 自体のサイズ変化を ResizeObserver で監視し、
-  // 子の `.xterm-viewport` を末尾までスクロールし直す。`.xterm-viewport` 自体は
-  // xterm.js が動的に生成するので querySelector で都度引く (mount/remount に追従)。
-  const termContainerRef = useRef<HTMLDivElement | null>(null);
+  const termRef = useRef<TerminalViewHandle | null>(null);
   const { settings } = useSettings();
   const t = useT();
   const confirmRemoveCard = useConfirmRemoveCard();
-  const setCardTitle = useCanvasStore((s) => s.setCardTitle);
   const setCardPayload = useCanvasStore((s) => s.setCardPayload);
   const { showToast } = useToast();
-  // Issue #253: Canvas zoom 下でも論理 px ベースで cols/rows を確定させる
-  const fit = useCanvasTerminalFit(settings);
   const payload = (data?.payload ?? {}) as AgentPayload;
   // 新スキーマ roleProfileId を優先、無ければ legacy role を読む
   const roleProfiles = useRoleProfiles();
@@ -186,77 +132,8 @@ function AgentNodeCardImpl({ id, data }: NodeProps): JSX.Element {
   const organizationAccent = visual.organizationAccent;
   const title = (data?.title as string) ?? visual.label;
   const [handoffBusy, setHandoffBusy] = useState(false);
-  // Issue #342 Phase 1: recruit 経路の spawn 失敗を Hub に ack するためのコールバック。
-  // payload.agentId / payload.teamId が揃っているとき (= 通常の AgentNode は常に揃う)
-  // のみ実体化し、それ以外は no-op を返す。
-  const onSpawnError = useRecruitSpawnAck(payload.agentId, payload.teamId);
   const [status, setStatus] = useState<string>('');
-  // Phase 4: ステータスバッジ。出力を最近受け取ったら typing、暫く来なければ idle。
   const [activity, setActivity] = useState<AgentStatus>('idle');
-  // Issue #125: 旧実装は 200ms 周期の setInterval を全 AgentNodeCard が常時動かしており
-  // 30 カード並ぶと idle 中も毎秒 150 回 timer が起きていた (省電力モード/裏画面でも)。
-  // → 出力イベント (handleActivity) の都度 setTimeout を立て直し、idle 復帰でクリアする。
-  //   typing 状態の間しかタイマーが動かないので idle 時はゼロコスト。
-  // useCallback で参照固定: TerminalView に渡す onActivity が毎レンダー新規になると、
-  //   TerminalView 内部の effect (handler 再 attach) が無駄に再実行される。
-  const idleTimerRef = useRef<number | null>(null);
-  const handleActivity = useCallback((): void => {
-    setActivity('typing');
-    if (idleTimerRef.current !== null) {
-      window.clearTimeout(idleTimerRef.current);
-    }
-    idleTimerRef.current = window.setTimeout(() => {
-      idleTimerRef.current = null;
-      setActivity((prev) => (prev !== 'idle' ? 'idle' : prev));
-    }, 600);
-  }, []);
-  useEffect(() => {
-    return () => {
-      if (idleTimerRef.current !== null) {
-        window.clearTimeout(idleTimerRef.current);
-        idleTimerRef.current = null;
-      }
-    };
-  }, []);
-
-  // ----- ユーザー入力から auto-summary タイトル -----
-  // 入力をバッファし、Enter (\r) を押した瞬間にバッファ内容をタイトル化する。
-  // 既にユーザーが手で名付けた (auto title 形式でない) 場合は上書きしない。
-  // title は user 入力で都度上書きされうるが、判定は確定タイミングだけなので ref で読む
-  // → callback 自体は (id, setCardTitle) のみに依存させて識別子を安定化。
-  const inputBufferRef = useRef('');
-  const titleRef = useRef(title);
-  titleRef.current = title;
-  const handleUserInput = useCallback(
-    (raw: string): void => {
-      if (!raw) return;
-      // 制御コードのうち BS, ESC, 矢印キー等は無視。Enter (\r/\n) は確定トリガ。
-      for (const ch of raw) {
-        const code = ch.charCodeAt(0);
-        if (ch === '\r' || ch === '\n') {
-          const text = inputBufferRef.current;
-          inputBufferRef.current = '';
-          const summary = summarizeInput(text);
-          if (summary && isAutoTitle(titleRef.current)) {
-            setCardTitle(id, summary);
-          }
-        } else if (code === 0x7f || code === 0x08) {
-          // Backspace
-          inputBufferRef.current = inputBufferRef.current.slice(0, -1);
-        } else if (code === 0x1b) {
-          // ESC シーケンス開始 → 同チャンク内の残りも捨てる近似
-          return;
-        } else if (code >= 0x20) {
-          inputBufferRef.current += ch;
-          // 暴走防止: 200 文字超えたらリセット
-          if (inputBufferRef.current.length > 200) {
-            inputBufferRef.current = inputBufferRef.current.slice(-200);
-          }
-        }
-      }
-    },
-    [id, setCardTitle]
-  );
 
   // Issue #23 + カスタムエージェント対応:
   // agent-resolver 経由で built-in (claude/codex) + customAgents のコマンド/引数/cwd を解決する。
@@ -288,7 +165,8 @@ function AgentNodeCardImpl({ id, data }: NodeProps): JSX.Element {
   });
   const teamMembers = useMemo(() => {
     if (!payload.teamId) return null;
-    if (teamMembersSig === '') return [] as { agentId: string; roleProfileId: string; agent: 'claude' | 'codex' }[];
+    if (teamMembersSig === '')
+      return [] as { agentId: string; roleProfileId: string; agent: 'claude' | 'codex' }[];
     return teamMembersSig.split(';').map((s) => {
       const [agentId, roleProfileId, agent] = s.split(':');
       return {
@@ -391,7 +269,7 @@ function AgentNodeCardImpl({ id, data }: NodeProps): JSX.Element {
       showToast(t('handoff.error.noProject'), { tone: 'error', duration: 8000 });
       return null;
     }
-    const snapshot = ref.current?.getBufferText(120) ?? '';
+    const snapshot = termRef.current?.getBufferText(120) ?? '';
     const kind = roleProfileId === 'leader' ? 'leader' : 'worker';
     const result = await window.api.handoffs.create({
       projectRoot,
@@ -464,7 +342,7 @@ function AgentNodeCardImpl({ id, data }: NodeProps): JSX.Element {
         // 全文が 1 つの paste として確定 → Claude/Codex が読み取って MCP を叩き始める。
         try {
           const prompt = buildLeaderHandoffPrompt(markdownPath, handoff.id);
-          ref.current?.sendCommand(wrapBracketedPaste(prompt), true);
+          termRef.current?.sendCommand(wrapBracketedPaste(prompt), true);
         } catch (err) {
           const detail = err instanceof Error ? err.message : String(err);
           showToast(t('handoff.error.injectFailed', { detail }), {
@@ -509,26 +387,6 @@ function AgentNodeCardImpl({ id, data }: NodeProps): JSX.Element {
         ['--organization-accent' as string]: organizationAccent ?? accent
       }) as React.CSSProperties,
     [accent, organizationAccent]
-  );
-
-  // Issue #261 / #272 / #272 v3: termContainer のサイズ変化時に xterm 自前の
-  // scroll model 経由で末尾までスクロールし直す。NodeResizer の縮小→拡大で
-  // scrollback 末尾が見切れるのを防ぐ。callback は xterm v6 の SmoothScrollableElement
-  // に正しく届くよう `Terminal.scrollToBottom()` を public API 経由で叩く
-  // (DOM の scrollTop 書換えは内部 scroll model と同期しないため使えない)。
-  const scrollToBottom = useCallback(() => {
-    ref.current?.scrollToBottom();
-  }, []);
-  useXtermScrollToBottomOnResize(termContainerRef, scrollToBottom);
-
-  // TerminalView に渡す onSessionId を useCallback 化。インライン lambda だと
-  // setActivity / status の局所 setState で AgentNodeCard が再描画するたび
-  // TerminalView 側の onSessionId が新規参照になり、内部 deps 経由の handler 差替えが起きる。
-  const handleSessionId = useCallback(
-    (sid: string): void => {
-      if (sid) setCardPayload(id, { resumeSessionId: sid });
-    },
-    [id, setCardPayload]
   );
 
   return (
@@ -589,48 +447,20 @@ function AgentNodeCardImpl({ id, data }: NodeProps): JSX.Element {
             </button>
           </span>
         </header>
-        <div
-          className="nodrag nowheel canvas-agent-card__term"
-          ref={termContainerRef}
-        >
-          <TerminalView
-            ref={ref}
-            // Issue #271: HMR remount で同じ PTY へ再 bind するための論理キー。
-            // ノード id は @xyflow/react canvas store で永続化されているので、
-            // HMR を跨いでも同一カードを一意に識別できる。
-            sessionKey={`canvas-agent:${id}`}
-            cwd={cwd}
-            fallbackCwd={cwd}
-            command={command}
-            // Issue #341: payload.args が空配列で永続化された場合に settings 由来の args が
-            // 潰れないようガード (`?? args` だと `[]` でも truthy 扱いで args が無視される)。
-            args={payload.args && payload.args.length > 0 ? payload.args : args}
-            codexInstructions={codexInstructions}
-            visible={true}
-            teamId={payload.teamId}
-            agentId={payload.agentId}
-            role={roleProfileId}
-            initialMessage={payload.initialMessage}
-            onStatus={setStatus}
-            onActivity={handleActivity}
-            onUserInput={handleUserInput}
-            onSessionId={handleSessionId}
-            // Issue #342 Phase 1: terminal_create 失敗を Hub に ack(false) する。
-            // 30 秒の handshake timeout を待たず、recruit MCP に構造化エラーが即返る。
-            onSpawnError={onSpawnError}
-            // Canvas zoom で xterm canvas が滲むのを避けるため WebGL を切る (DOM renderer 固定)。
-            // text は実 DOM になるので Chromium が親 transform に応じて再ラスタライズしシャープに描く。
-            disableWebgl
-            // Issue #272 v4: Canvas モードではホイールを scrollback スクロールへ強制ルーティング
-            // (xterm mouse protocol が wheel を消費して scrollback が動かない問題の対策)
-            forceWheelScrollback
-            // Issue #253: 論理 px ベース fit + zoom 購読 + 可観測性
-            unscaledFit={fit.unscaledFit}
-            getCellSize={fit.getCellSize}
-            zoomSubscribe={fit.zoomSubscribe}
-            getZoom={fit.getZoom}
-          />
-        </div>
+        <TerminalOverlay
+          cardId={id}
+          termRef={termRef}
+          payload={payload}
+          title={title}
+          roleProfileId={roleProfileId}
+          cwd={cwd}
+          command={command}
+          args={args}
+          codexInstructions={codexInstructions}
+          initialMessage={payload.initialMessage}
+          onStatus={setStatus}
+          onActivity={setActivity}
+        />
       </div>
       <Handle
         type="source"
